Refactor speed handling for player and enemies; adjust speed calculations based on game difficulty

+++ core/src/main/java/com/mathochiststudios/escapefromuni/GameDifficulty.java
@@ -21,8 +21,8 @@ public enum GameDifficulty {
         return baseMovementSpeed;
     }
 
-    public double getDeanSpeed() {
-        return deanSpeed;
+    public float getDeanSpeed() {
+        return (float) deanSpeed;
     }
 
     public double getSpeedBuffMultiplier() {
